Added priority customer logic and chart

+++ Realtime_Event_Ticketing_System_Back/src/main/java/com/w2053115/Sajitha/RealtimeTicketingSystem/service/implementation/CustomerServiceImpl.java
@@ -18,9 +18,16 @@ import java.util.ArrayList;
 public class CustomerServiceImpl implements CustomerService {
     private static final Logger logger = LoggerFactory.getLogger(CustomerServiceImpl.class);
 
+    //Normal Customer
     private final ArrayList<Thread> customerThreadList = new ArrayList<>();
     private final ArrayList<CustomerRunner> customerObjectList = new ArrayList<>();
+
+    //Priority Customer
+    private final ArrayList<Thread> priorityCustomerThreadList = new ArrayList<>();
+    private final ArrayList<CustomerRunner> priorityCustomerObjectList = new ArrayList<>();
+
     private int noOfCustomers = 0;
+    private int noOfPriorityCustomers = 0;
 
     @Autowired
     private CustomerRepo customerRepo;
@@ -32,11 +39,10 @@ public class CustomerServiceImpl implements CustomerService {
     private TicketPool ticketPool;
 
     @Override
-    public String createCustomer() {
-        if (configuration.getMaxTicketCapacity()  == 0 || configuration.getTotalTickets()  == 0 ||
-                configuration.getTicketReleaseRate()  == 0 || configuration.getCustomerRetrievalRate()  == 0) {
+    public void createCustomer(boolean priority) {
+        if (!isConfigurationLoaded()) {
             logger.warn("Load a Configuration before adding customers");
-            return "Load a Configuration before adding customers";
+            return;
         }
 
         //Add customer details to the database
@@ -45,7 +51,7 @@ public class CustomerServiceImpl implements CustomerService {
                 1,
                 "Gwen Stacy",
                 "0729950215",
-                1
+                priority
         );
         customerRepo.save(customer);
 
@@ -54,29 +60,67 @@ public class CustomerServiceImpl implements CustomerService {
         CustomerRunner customerObject = new CustomerRunner(
                 configuration.getCustomerRetrievalRate(),
                 1,
-                1,
+                priority,
                 ticketPool
         );
-        customerObjectList.add(customerObject);
+
+        //Adding to list depending on priority
+        if (priority) {
+            priorityCustomerObjectList.add(customerObject);
+        } else {
+            customerObjectList.add(customerObject);
+        }
 
         //Creating customer threads and storing it for accessing
         Thread customerThread = new Thread(customerObject);
-        customerThreadList.add(customerThread);
+
+        //Setting thread priority to 10 if priority customer
+        if (priority) {
+            customerThread.setPriority(Thread.MAX_PRIORITY);
+            priorityCustomerThreadList.add(customerThread);
+        } else { //Normally it's 5
+            customerThreadList.add(customerThread);
+        }
+
         if (SystemState.getState()==SystemState.RUNNING || SystemState.getState()==SystemState.PAUSED) {
             customerThread.start();
         }
-        for (Thread x : customerThreadList) {
-            System.out.println(x.getName());
+
+        if (priority) {
+            noOfPriorityCustomers++;
+            logger.info("[Priority] Customer {} created successfully", customerObject.getPriorityCustomerId());
+        } else {
+            noOfCustomers++;
+            logger.info("Customer {} created successfully", customerObject.getCustomerId());
         }
-        noOfCustomers++;
-        logger.info("Customer " + customerObject.getCustomerId() + " created successfully");
-        return "Customer " + customerObject.getCustomerId() + " created successfully";
     }
 
     @Override
-    public String removeCustomer() {
-        try {
-            if (customerRepo != null && !customerObjectList.isEmpty() && !customerThreadList.isEmpty()) {
+    public void removeCustomer(boolean priority) {
+        if (priority &&  customerRepo != null && !priorityCustomerObjectList.isEmpty() &&
+                !priorityCustomerThreadList.isEmpty()) {
+            try {
+
+                int priorityCustomerId = priorityCustomerObjectList.getLast().getPriorityCustomerId();
+
+                customerRepo.delete(customerRepo.findFirstByPriorityTrueOrderByCreatedDesc());
+
+                //Delete priority customer object and thread
+                priorityCustomerThreadList.getLast().interrupt();
+                priorityCustomerObjectList.removeLast();
+                priorityCustomerThreadList.removeLast();
+
+                noOfPriorityCustomers--;
+                CustomerRunner.reducePriorityId();
+                logger.info("[Priority] Customer {} Removed", priorityCustomerId);
+            } catch (NullPointerException e) {
+                logger.error("No records for priority customer", e);
+            }
+        }
+
+        if (!priority && customerRepo != null && !customerObjectList.isEmpty() && !customerThreadList.isEmpty()) {
+            try {
+
                 int customerId = customerObjectList.getLast().getCustomerId();
 
                 customerRepo.delete(customerRepo.findFirstByOrderByCreatedDesc());
@@ -88,20 +132,19 @@ public class CustomerServiceImpl implements CustomerService {
 
                 noOfCustomers--;
                 CustomerRunner.reduceId();
-                logger.info("Customer " + customerId + " Removed");
-                return "Customer " + customerId + " Removed";
-            } else {
-                logger.warn("There's no customers added");
-                return "There's no customers added";
+                logger.info("Customer {} Removed", customerId);
+
+            } catch (NullPointerException e) {
+                logger.error("No records for customer", e);
             }
-        } catch (NullPointerException e) {
-            logger.error("No records for customer");
-            throw e;
         }
     }
 
     @Override
     public void startCustomers() {
+        for (Thread thread : priorityCustomerThreadList) {
+            thread.start();
+        }
         for (Thread thread : customerThreadList) {
             thread.start();
         }
@@ -110,9 +153,7 @@ public class CustomerServiceImpl implements CustomerService {
 
     @Override
     public void stopCustomers() {
-        for (CustomerRunner customerRunner : customerObjectList) {
-            CustomerRunner.stop();
-        }
+        CustomerRunner.stop();
         logger.info("Customers stopped");
     }
 
@@ -127,18 +168,39 @@ public class CustomerServiceImpl implements CustomerService {
             for (Thread thread : customerThreadList) {
                 thread.interrupt();
             }
-            //Collections.fill(customerObjectList, null);
+
             customerObjectList.clear();
             customerThreadList.clear();
+        }
+        if (!priorityCustomerObjectList.isEmpty() && !priorityCustomerThreadList.isEmpty()) {
+            for (Thread thread : priorityCustomerThreadList) {
+                thread.interrupt();
+            }
+
+            priorityCustomerObjectList.clear();
+            priorityCustomerThreadList.clear();
         }
         CustomerRunner.resume();
         CustomerRunner.resetCustomerIds();
         noOfCustomers = 0;
+        noOfPriorityCustomers = 0;
         logger.info("Customers reset");
     }
 
     @Override
     public int getCustomers(){
         return noOfCustomers;
+    }
+
+    @Override
+    public int getPriorityCustomers(){
+        return noOfPriorityCustomers;
+    }
+
+    private boolean isConfigurationLoaded() {
+        return (configuration.getMaxTicketCapacity() > 0 &&
+                configuration.getTotalTickets() > 0 &&
+                configuration.getTicketReleaseRate() > 0 &&
+                configuration.getCustomerRetrievalRate() > 0);
     }
 }
